Add DetailsTile and use it to implement TrackScreenDetailsTile
