+++ packages/mobile/src/components/details-tile/index.ts
@@ -1,0 +1,1 @@
+export { DetailsTile } from './DetailsTile'
