+++ packages/mobile/src/components/details-tile/types.ts
@@ -1,6 +1,5 @@
 import { ReactNode } from 'react'
 
-import { PlaybackSource } from 'audius-client/src/common/models/Analytics'
 import { Collection } from 'audius-client/src/common/models/Collection'
 import { FavoriteType } from 'audius-client/src/common/models/Favorite'
 import { ID, UID } from 'audius-client/src/common/models/Identifiers'
@@ -10,46 +9,51 @@ import { RepostType } from 'audius-client/src/common/store/user-list/reposts/typ
 
 import { GestureResponderHandler } from 'app/types/gesture'
 
-export type LineupItemProps = {
-  /** Index of tile in lineup */
-  index: number
-
-  /** Are we in a trending lineup? Allows tiles to specialize their rendering */
-  isTrending?: boolean
-
-  /** Is this item unlisted (hidden)? */
-  isUnlisted?: boolean
-
-  /** Function to call when item & art has loaded */
-  onLoad?: (index: number) => void
-
-  /** Whether or not to show the artist pick indicators */
-  showArtistPick?: boolean
-
-  /** Whether to show an icon indicating rank in lineup */
-  showRankIcon?: boolean
-
-  /** Function that will toggle play of a track */
-  togglePlay: (uid: UID, id: ID, source: PlaybackSource) => void
-
+export type DetailsTileParentProps = {
   /** Uid of the item */
   uid: UID
 }
 
-export type LineupTileProps = Omit<LineupItemProps, 'togglePlay'> & {
-  children?: ReactNode
+export type DetailsTileDetail = {
+  icon?: ReactNode
+  isHidden?: boolean
+  label: string
+  value: ReactNode
+}
 
+export type DetailsTileProps = {
   /** Cosign information */
   coSign?: Track['_co_sign']
 
-  /** Duration of the tile's tracks */
-  duration?: number
+  /** Source for the analytics call when an external link in the description is pressed */
+  descriptionLinkPressSource: 'track page' | 'collection page'
+
+  /** Information about the item such as genre, duration, etc */
+  details: DetailsTileDetail[]
 
   /** Favorite type used for the favorited user list */
   favoriteType: FavoriteType
 
-  /** Hide the play count */
-  hidePlays?: boolean
+  /** Label to be displayed at the top of the tile */
+  headerText: string
+
+  /** Hide the favorite button */
+  hideFavorite?: boolean
+
+  /** Hide the favorite count */
+  hideFavoriteCount?: boolean
+
+  /** Hide the listen count */
+  hideListenCount?: boolean
+
+  /** Hide the overflow menu button */
+  hideOverflow?: boolean
+
+  /** Hide the repost button */
+  hideRepost?: boolean
+
+  /** Hide the repost count */
+  hideRepostCount?: boolean
 
   /** Hide the share button */
   hideShare?: boolean
@@ -63,11 +67,11 @@ export type LineupTileProps = Omit<LineupItemProps, 'togglePlay'> & {
   /** The item (track or collection) */
   item: Track | Collection
 
-  /** Function to call when tile is pressed */
-  onPress?: GestureResponderHandler
-
   /** Function to call when the overflow menu button is pressed */
   onPressOverflow?: GestureResponderHandler
+
+  /** Function to call when play button is pressed */
+  onPressPlay: GestureResponderHandler
 
   /** Function to call when repost is pressed */
   onPressRepost?: GestureResponderHandler
@@ -78,14 +82,17 @@ export type LineupTileProps = Omit<LineupItemProps, 'togglePlay'> & {
   /** Function to call when share is pressed */
   onPressShare?: GestureResponderHandler
 
-  /** Function to call when the title is pressed */
-  onPressTitle?: GestureResponderHandler
+  /** Render function for content below primary details */
+  renderBottomContent?: () => ReactNode
 
-  /** Amount of plays on this item */
-  playCount?: number
+  /** Render function for the header */
+  renderHeader?: () => ReactNode
 
   /** Repost type used for the reposted user list */
   repostType: RepostType
+
+  /** Amount of plays on this item */
+  playCount?: number
 
   /** Title of the item */
   title: string
